Move towards generic plotting
 - Move old plotting methods to legacy
 - Remove references to specifics e.g. DOS
 - Move DOS plotting to generic 1D plotting
