+++ setup.py
@@ -39,7 +39,7 @@ def run_setup(build_c=True):
         long_description = f.read()
 
     packages = ['euphonic',
-                'euphonic.plot',
+                'euphonic.legacy_plot',
                 'euphonic.readers']
 
     scripts = ['scripts/dispersion.py',
